Rename vm.c to liquid_vm.c
It's confusing to debug core dumps because Ruby also has vm.c so it
becomes hard to tell which frames come from Ruby and which ones
come from liquid-c.
